backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.36
pydantic==2.9.2
anthropic>=0.40.0
python-dotenv==1.0.1
